Update QuadTree.cpp
made quadrants boundaries int

+++ QuadTree/QuadTree.cpp
@@ -36,10 +36,10 @@ void QuadTree::subdivide() {
     if (point_count == 0) return; // No points to subdivide if none exist
 
     // Calculate the midpoints to divide the boundary into quadrants
-    const float midX = boundary.x;
-    const float midY = boundary.y;
-    const float halfWidth = boundary.w / 2;
-    const float halfHeight = boundary.h / 2;
+    const int midX = boundary.x;
+    const int midY = boundary.y;
+    const int halfWidth = boundary.w / 2;
+    const int halfHeight = boundary.h / 2;
 
     // Creates smaller boundary rectangles for each quadrant and initializes child QuadTrees
     northeast = std::make_unique<QuadTree>(Rect(midX + halfWidth, midY - halfHeight, halfWidth, halfHeight));
